+++ packages/jest-runtime/src/__tests__/runtime_require_module.test.js
@@ -351,6 +351,21 @@ describe('Runtime requireModule', () => {
       expect(exports.isJSONModuleEncodedInUTF8WithBOM).toBe(true);
     }));
 
+  it('should export a constructable Module class', () =>
+    createRuntime(__filename).then(runtime => {
+      const Module = runtime.requireModule(runtime.__mockRootPath, 'module');
+
+      expect(() => new Module()).not.toThrow();
+    }));
+
+  it('caches Module correctly', () =>
+    createRuntime(__filename).then(runtime => {
+      const Module1 = runtime.requireModule(runtime.__mockRootPath, 'module');
+      const Module2 = runtime.requireModule(runtime.__mockRootPath, 'module');
+
+      expect(Module1).toBe(Module2);
+    }));
+
   onNodeVersions('>=12.12.0', () => {
     it('overrides module.createRequire', () =>
       createRuntime(__filename).then(runtime => {