+++ scripts/buildTs.js
@@ -8,10 +8,13 @@
 'use strict';
 
 const fs = require('fs');
+const os = require('os');
 const path = require('path');
 
 const chalk = require('chalk');
 const execa = require('execa');
+const rimraf = require('rimraf');
+const throat = require('throat');
 const {getPackages} = require('./buildUtils');
 
 const packages = getPackages();
@@ -20,7 +23,13 @@ const packagesWithTs = packages.filter(p =>
   fs.existsSync(path.resolve(p, 'tsconfig.json'))
 );
 
-const args = ['tsc', '-b', ...packagesWithTs, ...process.argv.slice(2)];
+const args = [
+  '--silent',
+  'tsc',
+  '-b',
+  ...packagesWithTs,
+  ...process.argv.slice(2),
+];
 
 console.log(chalk.inverse(' Building TypeScript definition files '));
 
@@ -35,4 +44,55 @@ try {
   );
   console.error(e.stack);
   process.exitCode = 1;
+  return;
 }
+
+const downlevelArgs = ['--silent', 'downlevel-dts', 'build', 'build/ts3.4'];
+
+console.log(chalk.inverse(' Downleveling TypeScript definition files '));
+
+packagesWithTs.forEach(pkgDir => {
+  const pkg = require(pkgDir + '/package.json');
+
+  if (!pkg.types) {
+    throw new Error(`Package ${pkg.name} is missing \`types\` field`);
+  }
+
+  if (!pkg.typesVersions) {
+    throw new Error(`Package ${pkg.name} is missing \`typesVersions\` field`);
+  }
+
+  if (pkg.main.replace(/\.js$/, '.d.ts') !== pkg.types) {
+    throw new Error(
+      `\`main\` and \`types\` field of ${pkg.name} does not match`
+    );
+  }
+});
+
+// we want to limit the number of processes we spawn
+const cpus = Math.max(1, os.cpus().length - 1);
+
+Promise.all(
+  packagesWithTs.map(
+    throat(cpus, pkgDir => {
+      // otherwise we get nested `ts3.4` directories
+      rimraf.sync(path.resolve(pkgDir, 'build/ts3.4'));
+
+      return execa('yarn', downlevelArgs, {cwd: pkgDir, stdio: 'inherit'});
+    })
+  )
+)
+  .then(() => {
+    console.log(
+      chalk.inverse.green(
+        ' Successfully downleveled TypeScript definition files '
+      )
+    );
+  })
+  .catch(e => {
+    console.error(
+      chalk.inverse.red(' Unable to downlevel TypeScript definition files ')
+    );
+    console.error(e.stack);
+    process.exitCode = 1;
+  });